Add files via upload
In this project, we will be creating automation of 5 profiling tools (HPCToolkit, TAU, LIKWID, Intel VTune and Intel Offload Advisor) by creating shell scripts. We will be working on development of profiling tool on the basis of above tools. Final tool developed will be used to analyze the behaviour of any user application and find out its resource consumption.